Maybe fixed a bug
I think there was a bug with dropRows() that I fixed. I added some comments and removed dead code to make it more readable.

The way I originally wrote it made it confusing because I was subtracting 1 from both i and j but for different reasons. Subtracting 1 from i was a hack because there was padding on the width of the board. Subtracting 1 from j was necessary I think.