Still bug hunting
Last commit didn't fix the bug. The bug is that after a piece is placed that breaks rows a full row of empty blocks can sometimes appear.

There was a bug with scanCompletedRow() where the Board being passed to it was just a pointer to the first element in each column. Now it takes a pointer to the game instead.